scripts/ingest.py: Provide ingest reader
Well, I think this implemention is horrible because current implemention is not fit to the actual environment.

The real life is some APIs like the client make some HTTP requests to send messages to the server (Maybe the client is the comments server).

So it's impossible here to just have a single file...